models: Refactors error handling to be "module" specific
- Modifies consumed error handling from API server and controller modules
- Removes unused errors in `models/`
