+++ models/src/node/crd/v2.rs
@@ -3,7 +3,8 @@ use super::v1::BottlerocketShadowSpec as BottlerocketShadowSpecV1;
 use super::v1::BottlerocketShadowState as BottlerocketShadowStateV1;
 use super::v1::BottlerocketShadowStatus as BottlerocketShadowStatusV1;
 use super::BottlerocketShadowResource;
-use crate::node::{error, SEMVER_RE};
+use super::{error, Result};
+use crate::node::SEMVER_RE;
 
 use chrono::{DateTime, Utc};
 use kube::api::ObjectMeta;
@@ -181,7 +182,7 @@ impl BottlerocketShadowSpec {
 
     /// JsonSchema cannot appropriately handle DateTime objects. This accessor returns the transition timestamp
     /// as a DateTime.
-    pub fn state_timestamp(&self) -> error::Result<Option<DateTime<Utc>>> {
+    pub fn state_timestamp(&self) -> Result<Option<DateTime<Utc>>> {
         self.state_transition_timestamp
             .as_ref()
             .map(|ts_str| {
@@ -258,7 +259,7 @@ impl BottlerocketShadowStatus {
 
     /// JsonSchema cannot appropriately handle DateTime objects. This accessor returns the failure transition timestamp
     /// as a DateTime.
-    pub fn failure_timestamp(&self) -> error::Result<Option<DateTime<Utc>>> {
+    pub fn failure_timestamp(&self) -> Result<Option<DateTime<Utc>>> {
         self.state_transition_failure_timestamp
             .as_ref()
             .map(|ts_str| {
